requirements.txt
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
requests>=2.31.0
langchain>=0.1.0
langchain-openai>=0.0.5
